book 수정

+++ Book/src/main/java/onlinebookstore/OrderCancelled.java
@@ -3,7 +3,7 @@ package onlinebookstore;
 public class OrderCancelled extends AbstractEvent {
 
     private Long id;
-    private Integer bookId;
+    private Long bookId;
     private Integer qty;
     private Integer price;
     private Integer paymentId;
@@ -17,11 +17,11 @@ public class OrderCancelled extends AbstractEvent {
     public void setId(Long id) {
         this.id = id;
     }
-    public Integer getBookId() {
+    public Long getBookId() {
         return bookId;
     }
 
-    public void setBookId(Integer bookId) {
+    public void setBookId(Long bookId) {
         this.bookId = bookId;
     }
     public Integer getQty() {
